requirements.txt
# Core ML and Deep Learning
torch>=2.0.0
transformers>=4.35.0
datasets>=2.14.0
accelerate>=0.24.0
peft>=0.6.0

# Google AI for Teacher Model (Gemini)
google-generativeai>=0.3.0

# Evaluation and Metrics
scikit-learn>=1.3.0
numpy>=1.24.0
scipy>=1.11.0

# Data Processing
pandas>=2.0.0
tqdm>=4.65.0

# Configuration
pydantic>=2.0.0
pydantic-settings>=2.0.0
python-dotenv>=1.0.0

# Logging and Monitoring
wandb>=0.15.0
tensorboard>=2.14.0

# Utilities
pyyaml>=6.0
jsonlines>=3.1.0

# Optional: For serving
fastapi>=0.103.0
uvicorn>=0.23.0
